Adds resetting RTO after migration.
After migration, we want to reset the TCP RTO to a more sane value since
if we experience a long delay before reconnecting, we could end up missing
some connectivity. This can be common in cases of spotty Wifi. Need to
design a better answer than setting it to one, though.